add inline scripting feature for accessing out yaml tree
It is now possible to evaluate a property in the user's yaml by referencing a property in the processing output yaml.

The format for referencing a path in the output yaml is:

- {{ .property }}
- {{ .property[n] }}

And all the arbitrary compositions of the above base cases.
An abbreviated format is also possible:

{{ .[n][m].property }}

This is equivalent to:

{{ .conversations[n].requests[m].property }}

Example:

conversations:
    host: localhost:7480
    requests:
      - method: POST
        uri: test-5/mp.3
        queryString: format=json&uploads
      - method: PUT
        uri: test-5/mp.3
        queryString: format=json&partNumber=1&uploadId={{.[0][0].response.body.UploadId}}

The queryString property in the second request is evaluated referencing the 'response.body.UploadId' path starting from the .conversations[0].requests[0] node in the output yaml.

Signed-off-by: Giuseppe Baccini <giuseppe.baccini@suse.com>